Fixed Michael beating the AI (for now) See Notes
In the way that Michael won the game, the AI thought that if Michael were to make the best move he can for the next 6 moves, it would lose. Does this prove that the AI is beatable? Maybe, we'll have to see...

In order to remedy the situation, when the AI has all of its possible moves with a score of -infinity, it will try to reduce its depth level to make the best move it possibly can in that scenario

This is the state of the game that Michael won on:

0 0
W W W W W
W W W W W
W W W W k
W W W k W
W W W W K

You used to be able to win by moving the knight at row 3 column 5 to row 4 column 3. Try to beat it again Michael!